[ADD] Done. Implement MyString class

+++ CPP/HongLab/ch03Array/StringADT/MyString.h
@@ -47,7 +47,7 @@ public:
 	size_t	Length();
 	void	Resize(size_t new_size);
 
-	MyString Substr(int start, int num);	// 인덱스 start위치의 글자부터 num개의 글자로 새로운 문자열 만들기
+	MyString Substr(size_t start, size_t num);	// 인덱스 start위치의 글자부터 num개의 글자로 새로운 문자열 만들기
 	MyString Concat(MyString app_str);		// 뒤에 덧붙인 새로운 문자열 반환 (append)
 	MyString Insert(MyString target, size_t start); // 중간에 삽입
